requirements.txt
accelerate
codetiming
datasets
flash-attn==2.7.4.post1
liger-kernel
mathruler
numpy
omegaconf
pandas
peft
pillow
pyarrow>=15.0.0
pylatexenc
qwen-vl-utils==0.0.14  # Required for Qwen3-VL
scipy
matplotlib
ipykernel
# sam2  # Only needed for segmentation evaluation (bbox/point → mask), not for existence detection training
# pycocotools  # Only needed for segmentation evaluation, not for existence detection training
scikit-image
ray
tensordict
# Updated for Qwen3-VL-8B support
transformers>=4.57.0  # Qwen3-VL requires transformers>=4.57.0
vllm>=0.11.0  # Qwen3-VL requires vllm>=0.11.0
wandb
